SAAS-1139 migration to mui (#328)
* SAAS-1162: MUI PrivateLayout redesign (#309)

* SAAS-1162: New private layout design changes

* SAAS-1162: Add keys to side menu

* SAAS-1162: Add test-id and check for resources

* SAAS-1162: Rename MenuBar to Header and add consts

* SAAS-1162: Fix Playwright tests

* SAAS-1162: Rename Header to AppHeader
Fix issues from CR

* SAAS-1162: import fonts and split NavItem

* SAAS-1162: Split Nav items into separate files

* SAAS-1162: Add back incorrect route

* SAAS-7 Updates readme to reflect change in testing framework (#313)

* Fix for the UI tests

* SAAS-1162: Put modals over header and side menu

* SAAS-1162: Put profile menu over toasts


* SAAS-1158: mui integration profile (#317)

* SAAS-1158 add mui components

* SAAS-1158 add data test id

* SAAS-1158 fix data ids for tests

* SAAS-1158 fix env

* SAAS-1158 add validation

* SAAS-1158 finish text input

* Fix for the UI tests

* SAAS-1158 remove ts-ignores

* SAAS-1158 add test to textfield

* SAAS-1158 remove validation function

* Fix the UI tests

* SAAS-1158 add refactoring

* SAAS-1158 add emotion styled


* SAAS-1159: mui integration 404 page (#322)

* SAAS-1159 remove grafana ui

* SAAS-1159 add text transform to button

* SAAS-1159 add dependency

* SAAS-1159 remove useless packets from package

* Fix for the UI tests


* SAAS-1153 migrate tables to MUI (#316)

* SAAS-1153 add Overlay

* SAAS-1153 update .gitignore

* SAAS-1153 add dependencies

* SAAS-1153 add style utils

* SAAS-1153 use MUI's DateGrid instead of Table

* SAAS-1153 update MembersList

* SAAS-1153 use MUI's DataGrid instead of Table

* SAAS-1153 use className instead of css

* Fix the UI tests

* Fix unused variable

* SAAS-1153 update dependencies

* SAAS-1153 add SimpleDialog

* SAAS-1153 use SimpleDialog instead of Modal

* SAAS-1153 fix typos in PmmInstanceRemoveModal

* SAAS-1153 use SimpleDialog in MemberDeleteModal

* SAAS-1153 minor change

* SAAS-1153 add mui-rff dependencies

* SAAS-1153 update SimpleDialog

* SAAS-1153 use SimpleDialog instead of Modal in MemberDeleteModal

* SAAS-1153 minor refactoring

* Squashed commit of the following:

commit 200671edac42f788202a0c024707e9031e09476f
Author: Tomáš Blažek <100355118+TomasBlazek3PG@users.noreply.github.com>
Date:   Wed Nov 2 14:54:50 2022 +0100

    SAAS-1162: MUI PrivateLayout redesign (#309)

    * SAAS-1162: New private layout design changes

    * SAAS-1162: Add keys to side menu

    * SAAS-1162: Add test-id and check for resources

    * SAAS-1162: Rename MenuBar to Header and add consts

    * SAAS-1162: Fix Playwright tests

    * SAAS-1162: Rename Header to AppHeader
    Fix issues from CR

    * SAAS-1162: import fonts and split NavItem

    * SAAS-1162: Split Nav items into separate files

    * SAAS-1162: Add back incorrect route

    * SAAS-7 Updates readme to reflect change in testing framework (#313)

    * Fix for the UI tests

    * SAAS-1162: Put modals over header and side menu

    * SAAS-1162: Put profile menu over toasts


* SAAS-1153 use DataGrid instead of Table in TicketList

* SAAS-1153 fix table spacing

* SAAS-1153 add custom empty table message

* SAAS-1153 remove table footer in PmmInstanceList

* SAAS-1153 add tests for SimpleDialog

* SAAS-1153 add tests for MemberDeleteModal

* SAAS-1153 minor update

* SAAS-1153 add tests for MemberEditModal

* SAAS-1153 update tests

* Part of the fix for the UI tests

* Part of fix for the UI tests

* SAAS-1153 fix pagination

* SAAS-1153 fix linting

* SAAS-1153 fix pagination

* Fix for the UI tests

* Change element in UI tests

* Change of element in UI tests

* Scrolls to element in UI tests

* Split UI pagination test

* Change locator in UI tests for pagination select


* SAAS-1139: Fix merge issues

* SAAS-1139 update package-lock.json to new version

* SAAS-1161 migrate the login page to MUI (#331)

* SAAS-1161 migrate logger from core

* SAAS-1161 migrate Login to MUI

* SAAS-1161 add a basic test for Login

* SAAS-1161 refactoring of logger

* SAAS-1161 use the Typography component

* SAAS-1157 migrate the PMM instances list to MUI (#330)

* SAAS-1157 migrate PmmInstanceActions to MUI

* SAAS-1157 migrate ManagePmmInstance to MUI

* SAAS-1157 minor fix

* SAAS-1157 add tests for ManagePmmInstances

* SAAS-1157 fix ManagePmmInstances.tsx

* SAAS-1157 fix header

* SAAS-1156: mui integration kubernetes (#326)

* SAAS-1156 change kunernetes components

* SAAS-1156 fix styles modal

* SAAS-1156 add react content loader

* SAAS-1156 remove background on modal


* SAAS-1160: Activation Page MUI Migration (#327)

* SAAS-1160: Migrate fields and styles

* SAAS-1160 fix core/utils.ts

* SAAS-1160: Fix tests

* SAAS-1160: Change button to contained version

* SAAS-1160: Use Icons and use predefined colors


* SAAS-1190 Automate Tests SAAS-T218, SAAS-T229, SAAS-T232, SAAS-T250 (#319)

* Adds test SAAS-T232

* Adds test SAAS-T218

* Adds test SAAS-T250

* adds test 'SAAS-T229

* Test cases fix

* SAAS-7 Longer timeout for the tests (#339)

* SAAS-1127 Empty constructor for Page Objects and refactoring (#333)

* SAAS-1127: Creating Empty constructor for the page objects

* SAAS-1127: Sign Up page refactor

* SAAS-1127: Refactors rest of page objects

* Fixing the merge issues

* Remove ThemeContext from Grafana (#337)

* SAAS-1155: Migrate Dashboard to MUI (#338)

* SAAS-1155: Migrate:
- Contacts
- PromoBanner
- Dashboard
Remove:
-IconLink

* SAAS-1155: Fix EntitlementModal

* SAAS-1155: Add expiry date

* SAAS-1155: Refactor expiration date


* SAAS-1211 migrate getting started to MUI (#335)

* SAAS-1211 migrate GettingStarted to MUI

* SAAS-1211 migrate GettingStartedSection to MUI

* SAAS-1211 fix add button icon

* SAAS-1211 update styles

* SAAS-1211 fix Link

* SAAS-1154 migrate ManageOrganization to MUI (#329)

* SAAS-1154 migrate ManageOrganization to MUI

* SAAS-1154 update style

* SAAS-1154 migrate OrganizationView to MUI

* SAAS-1154 fix organization name in deletion modal

* SAAS-1154 use LoadingButton, Button and Alert from MUI

* SAAS-1154 migrate OrganizationDeleteModal to MUI

* SAAS-1154 migrate ResendEmailLink to MUI

* SAAS-1154 fix pending invitation style

* SAAS-1154 migrate MemberActions to MUI

* SAAS-1154 fix tests for OrganizationDeleteModal

* SAAS-1154 use IconButton from MUI

* SAAS-1154 migrate OrganizationCreate to MUI

* SAAS-1154 migrate OrganizationCreate to MUI

* SAAS-1154 refactoring of OrganizationCreate.styles.ts

* SAAS-1154 fix org edit validation and tests

* SAAS-1154 fix default tab

* SAAS-1154 use rff-mui

* SAAS-1154: Fix validation schema

* SAAS-1154 fix style

* SAAS-1154 use MUI's Dialog instead of our Modal

* SAAS-1154 fix org deletion

* SAAS-1154 update messages

* SAAS-1154 minor refactoring


* SAAS-1207: Migrate Ticket Overview to MUI (#334)

* SAAS-1207: Migrate to MUI

* SAAS-1207: Remove black border


* SAAS-1176: Handle Invalid Activation Links (#332)

* SAAS-1176: Add content loader and swap of tokens

* SAAS-1176:
- Add Token Validation
- Add Error Page
- Update Tests

* SAAS-1176: Five minutes TTL to constants

* SAAS-1176 update invalid token style


* SAAS-1139 fix InviteMember

* SAAS-1139 Fix UI tests (#340)

* Fix UI tests

* Fix invite member

* Fix for the invite user

* Changes method in UI tests

+++ ui-tests/tests/auth/login.spec.ts
@@ -8,17 +8,16 @@ import { getUser } from '@helpers/portalHelper';
 test.describe('Spec file for dashboard tests for customers', async () => {
   let adminUser: User;
 
-  test.beforeAll(async () => {
+  test.beforeEach(async ({ page }) => {
     adminUser = getUser();
     await oktaAPI.createUser(adminUser, true);
-  });
-
-  test.afterAll(async () => {
-    await oktaAPI.deleteUserByEmail(adminUser.email);
-  });
-
-  test.beforeEach(async ({ page }) => {
     await page.goto('/');
+  });
+
+  test.afterEach(async () => {
+    if (await oktaAPI.getUser(adminUser.email)) {
+      await oktaAPI.deleteUserByEmail(adminUser.email);
+    }
   });
 
   test('SAAS-T251 - Verify handling of status 401 @auth', async ({ page }) => {
@@ -47,46 +46,49 @@ test.describe('Spec file for dashboard tests for customers', async () => {
     await signInPage.waitForPortalLoaded();
     expect(page.url()).toContain(baseURL);
 
-    await dashboardPage.locators.gettingStartedContainer.waitFor({ state: 'visible' });
+    await dashboardPage.gettingStartedContainer.waitFor({ state: 'visible' });
   });
 
-  test('SAAS-T86 Verify unsuccessful login on Percona Portal @login @auth', async ({ page, context }) => {
-    const signInPage = new SignInPage(page);
-    const dashboardPage = new DashboardPage(page);
+  if (process.env.CI) {
+    // Test Fails on local, runs only in CI.
+    test('SAAS-T86 Verify unsuccessful login on Percona Portal @login @auth', async ({ page, context }) => {
+      const signInPage = new SignInPage(page);
+      const dashboardPage = new DashboardPage(page);
 
-    await test.step('SAAS-T245 Verify user is able to see "Continue with.." buttons', async () => {
-      await expect(signInPage.continueGoogle).toHaveText(signInPage.continueGoogleLabel);
-      await expect(signInPage.continueGitHub).toHaveText(signInPage.continueGitHubLabel);
+      await test.step('SAAS-T245 Verify user is able to see "Continue with.." buttons', async () => {
+        await expect(signInPage.continueGoogle).toHaveText(signInPage.continueGoogleLabel);
+        await expect(signInPage.continueGitHub).toHaveText(signInPage.continueGitHubLabel);
 
-      const [googlePage] = await Promise.all([
-        context.waitForEvent('page'),
-        signInPage.continueGoogle.click(),
-      ]);
-      const googleSigInPage = new SignInPage(googlePage);
+        const [googlePage] = await Promise.all([
+          context.waitForEvent('page'),
+          signInPage.continueGoogle.click(),
+        ]);
+        const googleSigInPage = new SignInPage(googlePage);
 
-      await googleSigInPage.googleEmailField.waitFor({ state: 'visible', timeout: 60000 });
+        await googleSigInPage.googleEmailField.waitFor({ state: 'visible', timeout: 60000 });
 
-      expect(googlePage.url()).toContain('https://accounts.google.com/o/oauth2/auth/');
-      await googlePage.close();
+        expect(googlePage.url()).toContain('https://accounts.google.com/o/oauth2/auth/');
+        await googlePage.close();
 
-      const [gitHubPage] = await Promise.all([
-        context.waitForEvent('page'),
-        signInPage.continueGitHub.click(),
-      ]);
-      const gitHubSigInPage = new SignInPage(gitHubPage);
+        const [gitHubPage] = await Promise.all([
+          context.waitForEvent('page'),
+          signInPage.continueGitHub.click(),
+        ]);
+        const gitHubSigInPage = new SignInPage(gitHubPage);
 
-      await gitHubSigInPage.gitHubEmailField.waitFor({ state: 'visible', timeout: 60000 });
+        await gitHubSigInPage.gitHubEmailField.waitFor({ state: 'visible', timeout: 60000 });
 
-      expect(gitHubPage.url()).toContain('https://github.com/login');
-      await gitHubPage.close();
+        expect(gitHubPage.url()).toContain('https://github.com/login');
+        await gitHubPage.close();
+      });
+
+      await signInPage.fillOutSignInUserDetails('Wrong Username', 'WrongPassword');
+      await signInPage.signInButton.click();
+      await signInPage.signInErrorContainer.waitFor({ state: 'visible' });
+      await expect(signInPage.signInErrorContainer).toHaveText(signInPage.unableToSignIn);
+      expect(page.url()).toContain(dashboardPage.routes.login);
+
+      await dashboardPage.gettingStartedContainer.waitFor({ state: 'detached' });
     });
-
-    await signInPage.fillOutSignInUserDetails('Wrong Username', 'WrongPassword');
-    await signInPage.signInButton.click();
-    await signInPage.signInErrorContainer.waitFor({ state: 'visible' });
-    await expect(signInPage.signInErrorContainer).toHaveText(signInPage.unableToSignIn);
-    expect(page.url()).toContain(dashboardPage.routes.login);
-
-    await dashboardPage.locators.gettingStartedContainer.waitFor({ state: 'detached' });
-  });
+  }
 });
